Implement ORM layer, with skeleton repository structure, also tweak domain models to fulfill ef core structure requirements

+++ TeslaMall.Server/Services/Contracts/IRentCalculatorService.cs
@@ -4,5 +4,5 @@ namespace TeslaMall.Server.Services.Contracts;
 
 public interface IRentCalculatorService
 {
-    public decimal CalculateCostOfRent(Reservation reservation);
+    public float CalculateCostOfRent(Reservation reservation);
 }
